button component to base buttoncomponent, add regis page
