+++ src/app/(user)/page.tsx
@@ -1,13 +1,28 @@
 import React from "react";
 import styles from "./page.module.css";
-import Button from "@/components/user/button";
+import Button from "@/components/user/Button";
 
 const UserLoginPage = () => {
   return (
     <div className="">
-      <p className="text-xl font-extrabold divide-y-2">오마타세</p>
+      <header className="text-2xl font-extrabold">오마타세</header>
       <hr className={styles.force_ignore_padding} />
-      <Button color="primary">시작하기</Button>
+      <section>
+        <div>서비스 이용이 처음이라면</div>
+      </section>
+      <section>
+        <div>일본 음식점 예약 이제 줄 서지 마세요</div>
+      </section>
+      <section>
+        첫 번째 예약
+        <input type="text" />
+        <input type="text" />
+      </section>
+      <section>
+        <Button size="sm" color="primary">
+          시작하기
+        </Button>
+      </section>
     </div>
   );
 };
