+++ src/app/(user)/registration/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const RegistrationPage = () => {
+  return <div>RegistrationPage</div>;
+};
+
+export default RegistrationPage;
