+++ src/app/guide/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Button from "@/components/user/button";
+import Button from "@/components/common/BaseButton";
 import React from "react";
 import "@root/globals.css";
