+++ src/app/types/button/index.ts
@@ -1,0 +1,7 @@
+export interface UserButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  children: React.ReactNode;
+  isRadius?: boolean;
+  color: "primary" | "kakao";
+  size: "sm" | "lg" | "full";
+}
