+++ src/components/user/button.tsx
@@ -1,55 +1,18 @@
+"use client";
+
 import React from "react";
-import cls from "classnames";
-/*
-cls 사용이유
-- isRadis ? "rounded-lg":"" => 이런 것들을 자동으로 없애줌
-- trim()같은 메서드 없이 자동으로 문자열을 합쳐주지 때문에 가독성을 높일 수 있다.
-*/
+import BaseButton from "../common/BaseButton";
 
-interface UserButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  children: React.ReactNode;
-  isRadius?: boolean;
-  color: "primary" | "kakao";
-  size: "sm" | "lg" | "full";
-}
+import { UserButtonProps } from "@/app/types/button";
 
-const btnSize = {
-  sm: "w-[309px]",
-  lg: "w-[340px]",
-  full: "w-full",
-};
-
-const bgColorMap: Record<UserButtonProps["color"], string> = {
-  primary: "bg-primary",
-  kakao: "bg-kakao",
-};
-
-const textColorMap: Record<UserButtonProps["color"], string> = {
-  primary: "text-white",
-  kakao: "text-black",
-};
-
-const Button = ({
-  children,
-  isRadius = false,
-  color,
-  size,
-  className,
-  ...props
-}: UserButtonProps) => {
-  const btnClassNames = cls(
-    bgColorMap[color],
-    textColorMap[color],
-    btnSize[size],
-    { "rounded-lg": isRadius },
-    "h-[50px]",
-    className
-  );
+const Button = ({ children, ...rest }: UserButtonProps) => {
+  const handleClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    console.log(e.target);
+  };
   return (
-    <button className={btnClassNames} {...props}>
+    <BaseButton onClick={(e) => handleClick(e)} {...rest}>
       {children}
-    </button>
+    </BaseButton>
   );
 };
